Abstract filter logic from PaginatedLoader to PaginatedLoaderFiltered
