+++ src/utils/Helper.ts
@@ -43,3 +43,9 @@ export function getPropertyByPath(obj, path) {
         return prev ? prev[curr] : null
     }, obj || self);
 }
+
+export function indexOfEnd(baseString: string, searchString: string) {
+    const index = baseString.indexOf(searchString);
+    
+    return index == -1 ? -1 : index + searchString.length;
+}
